fix: refactor to not use partially-applied function api
there seems to be an issue with the way TypeScript handles the types when converting from a function of an ArgumentList to a function that's curried based on the number of arguments.

this style of function composition is useful for creating functions from functions, which was the original design of the hook implementation for `useReaction`. However, there seemed to be an issue with how partially applying the function erased some type information about the function being curried. So in effect, a clear type argument would become unknown because the function was curried from a non-curried function.

so to resolve the issue, I redefined the hook implementation to generate a function that creates an event handler from a single function call instead of multiple functions calls through currying. in the end, this next approach seems simpler and more concise, and with less custom memoization happening.